Removed GeoTools, ingest works with native GeoTiff reader.
